Support mcpServers format
